+++ checks_other.go
@@ -19,3 +19,5 @@ func platformChecks() []Check {
 
 func fontDirs() []string           { return nil }
 func firefoxProfileDirs() []string { return nil }
+
+func chromiumBrowsers() []chromiumBrowser { return nil }